+++ test/suite/lang.java.test.ts
@@ -30,7 +30,8 @@ suite("Java Language Test Suite", () => {
     if ( ! m ) ;
     }
     }
-    } <ENTER>`;
+    } <ENTER>
+    @ TestClass SS { }`;
     const sourceExpect = `import java.util.regex.Matcher;
     import java.util.regex.Pattern;
     public class RegexMatches {
@@ -48,7 +49,8 @@ suite("Java Language Test Suite", () => {
                 if (!m);
             }
         }
-    }↵`;
+    }↵
+    @TestClass SS { }`;
 
     const sourceTokensLines = sourceTokens.split("\n").map((_) => _.trim());
     const sourceExpectLines = sourceExpect.split("\n").map((_) => _.trim().replace("↵", "\n"));