+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Change Log
 All notable changes to the "aiXcoder" extension will be documented in this file.
 
+## [0.1.18] - %DATE%
+
+### Added
+### Changed
+- Windows 7 及之前版本中，预测前面的符号默认为★。因为Emoji⭐会显示成方块。Mac不受影响。
+
+### Fixed
+- 去掉了@后面的多余的空格。
+- 在未保存的c/c++文件中也有预测了。
+
 ## [0.1.17] - 2019-09-27
 
 ### Added